+++ cli/src/utils/mod.rs
@@ -1,0 +1,5 @@
+pub mod fs;
+pub mod mime;
+
+pub use fs::*;
+pub use mime::*;
