+++ client/src/tape/write.rs
@@ -26,8 +26,7 @@ pub async fn write_to_tape(
     );
 
     let sig = send_with_retry(client, &instruction, signer, MAX_RETRIES).await?;
-    println!("DEBUG: {}", sig);
-
+    //println!("DEBUG: {}", sig);
     Ok(sig)
 }
